requirements.txt
requests
slackclient
PyYAML
emoji
sqlalchemy
pandas
